refactor: adding folders to separate components
