+++ front/src/components/session/EditSessionModal.js
@@ -1,10 +1,10 @@
 import { useEffect, useState } from "react";
-import Modal from "./Modal";
-import useAxiosPrivate from "../hooks/useAxiosPrivate";
-import useExercise from "../hooks/useExercise";
+import Modal from "../Modal";
+import useAxiosPrivate from "../../hooks/useAxiosPrivate";
+import useExercise from "../../hooks/useExercise";
 import { useLocation, useNavigate } from "react-router-dom";
-import useSession from "../hooks/useSession";
-import { formatDateToYMD } from "../utils/dateUtils";
+import useSession from "../../hooks/useSession";
+import { formatDateToYMD } from "../../utils/dateUtils";
 
 const SESSIONS_URL = "/session";
